+++ framework/math/SpatialDiscretization/FiniteVolume/FiniteVolume.h
@@ -5,11 +5,13 @@
 
 #include <map>
 
-//###################################################################
 namespace chi_math::spatial_discretization
 {
-/**Spatial discretizations supporting Finite Volume representations.
-\ingroup doc_SpatialDiscretization*/
+
+/**
+ * Spatial discretizations supporting Finite Volume representations.
+ * \ingroup doc_SpatialDiscretization
+ */
 class FiniteVolume : public SpatialDiscretization
 {
 private:
@@ -20,25 +22,27 @@ private:
 
 public:
   virtual ~FiniteVolume() = default;
-  // prevent anything else other than a shared pointer
+
+  /**
+   * Publicly accessible construction handler.
+   */
   static std::shared_ptr<FiniteVolume>
   New(const chi_mesh::MeshContinuum& in_grid,
       CoordinateSystemType in_cs_type = CoordinateSystemType::CARTESIAN);
 
-  // 01
   void CreateCellMappings();
 
-  // 02 node ordering
 protected:
+  /**
+   * Develops node ordering per location.
+   */
   void OrderNodes();
 
-  // 03 sparsity
 public:
   void BuildSparsityPattern(std::vector<int64_t>& nodal_nnz_in_diag,
                             std::vector<int64_t>& nodal_nnz_off_diag,
                             const UnknownManager& unknown_manager) const override;
 
-  // 04a mappings
   int64_t MapDOF(const chi_mesh::Cell& cell,
                  unsigned int node,
                  const UnknownManager& unknown_manager,
@@ -55,12 +59,12 @@ public:
   {
     return MapDOF(cell, node, UNITARY_UNKNOWN_MANAGER, 0, 0);
   }
+
   int64_t MapDOFLocal(const chi_mesh::Cell& cell, unsigned int node) const override
   {
     return MapDOFLocal(cell, node, UNITARY_UNKNOWN_MANAGER, 0, 0);
   }
 
-  // 05 utils
   size_t GetNumGhostDOFs(const UnknownManager& unknown_manager) const override;
   std::vector<int64_t> GetGhostDOFIndices(const UnknownManager& unknown_manager) const override;
 };
